separate common methods into elasticsearch specific and shared methods
specific elasticsearch methods are now into the main elasticsearch
class file and shared methods have been moved into the
PluginMixins::ElasticSearch::Common namespace.
The license checking code has been externalized and can now be
specified as argument to the build_client method.
This is code refactorting that has no end-user impact.
