+++ lib/logstash/plugin_mixins/elasticsearch/noop_license_checker.rb
@@ -1,0 +1,9 @@
+module LogStash; module PluginMixins; module ElasticSearch
+  class NoopLicenseChecker
+    INSTANCE = self.new
+
+    def appropriate_license?(pool, url)
+      true
+    end
+  end
+end; end; end
